Make env_destroy_all ignore active containers.

+++ env_destroy_all.sh
@@ -13,17 +13,37 @@ unmount_chroot()
     | sed -nre 's@^('"$escaped"'/.*)$@\1@gp')
 }
 
+active_lxc_roots=()
+
 while read -r n
 do
-  echo "Destroying LXC container $n"
+  echo "Finding root for LXC container $n"
   t="$(lxc-info --name="${n}" -c lxc.rootfs \
     | sed -nre 's@^\S+\s*=\s*(\S+)$@\1@gp')"
+
+  echo "Checking if LXC container $n is up"
+  if (( "$(lxc-ls -a | sed -nre 's@^('"$n"')$@\1@gp' | wc -l)" > "0" ))
+  then
+    echo "Ignoring active container $n"
+    active_lxc_roots+=("$t")
+    continue
+  fi
+
+  echo "Destroying LXC container $n"
   unmount_chroot "$t"
   lxc-destroy --name="$n"
 done < <(lxc-ls -l | awk '{print $9}')
 
 for t in $(find /tmp -mindepth 1 -maxdepth 1 -iname 'tmp.*')
 do
+  for root in "${active_lxc_roots[@]}"
+  do
+    if [[ "$root" == "$t" ]]
+    then
+      echo "Ignoring root $t owned by active LXC session"
+      continue 2
+    fi
+  done
   echo "Destroying /tmp chroot $t..."
   unmount_chroot "$t"
   rm -rf "$t"
